MAINT: address reviewer comments
- employ use of random_state where appropriate
- move test to test_workflow
- fix mypy type hint
