+++ viral_seq/tests/test_feature_importance.py
@@ -8,7 +8,6 @@ from importlib.resources import files
 import shap
 import pytest
 import pandas as pd
-from sklearn.ensemble import RandomForestClassifier
 
 
 matplotlib.use("Agg")
@@ -217,7 +216,12 @@ def test_plot_shap_consensus(tmp_path):
     max_features = len(syn_data)
 
     fi.plot_shap_consensus(
-        (syn_shap_values, syn_data), syn_df, "Test", max_features, tmp_path, rng=rng
+        (syn_shap_values, syn_data),
+        syn_df,
+        "Test",
+        max_features,
+        tmp_path,
+        random_state=123,
     )
     assert (
         compare_images(
@@ -227,39 +231,3 @@ def test_plot_shap_consensus(tmp_path):
         )
         is None
     )
-
-
-def test_pearson_aggregation():
-    # enforce Pearson aggregation behavior, avoid reduction across folds
-    random_state = 123
-    rng = np.random.default_rng(random_state)
-    kmer_data = rng.integers(0, 2, size=(1000, 12))
-    data_target = np.asarray([1, 0] * 500)
-    data_target[-1] = 1
-
-    kmer_names = np.array([f"kmer_{i}" for i in range(12)])
-
-    train_data = pd.DataFrame(kmer_data, columns=kmer_names)
-    y = pd.Series(data_target)
-
-    classifier_parameters = dict(
-        clfr_name=RandomForestClassifier, n_estimators=100, n_jobs=None
-    )
-    (feature_count, shap_clfr_consensus, clfr_preds) = workflow.train_clfr(
-        train_data,
-        y,
-        classifier_parameters,
-        n_folds=2,
-        max_features=3,
-        random_state=random_state,
-    )
-
-    pearson_rank = feature_count["Pearson R"]
-    pearson_rank_exp = [
-        -0.11839685961062657,
-        -0.19111178312660876,
-        0.1441065507615085,
-        0.10391636402704744,
-    ]
-    # check first four pearson values, numbers have tendency to vary slightly based on dependency versions
-    assert_allclose(pearson_rank[: len(pearson_rank_exp)], pearson_rank_exp)
